# [WIP] Change EvalDoubleVisitor to use BaseVisitor