This PR is to see whether changing from Visitor to BaseVisitor will affect the performance of EvalDoubleVisitor. If not, I'll change all the Visitors to BaseVisitor.

For me, I noticed no slowdown using gcc.

@certik