@thilinarmtb 2015-04-27
master:
```
thilina@thilina-Inspiron:csympy$ benchmarks/eval_double1 
8362ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
thilina@thilina-Inspiron:csympy$ benchmarks/eval_double1 
8457ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
thilina@thilina-Inspiron:csympy$ benchmarks/eval_double1 
8300ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
```

this branch:
```
thilina@thilina-Inspiron:csympy$ benchmarks/eval_double1 
8293ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
thilina@thilina-Inspiron:csympy$ benchmarks/eval_double1 
8184ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
thilina@thilina-Inspiron:csympy$ benchmarks/eval_double1 
8613ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
```