@abinashmeher999 2015-05-01
@isuruf Is 'visitor' here same as 'visitor' in the Visitor Design Pattern?