@isuruf 2015-05-02
@certik, I removed the methods that were throwing not implemented errors into one method for Basic, but I think it's better to keep them there so that we do not miss some Basic classes when we extend the functionality to include `Log`, `ASinh` etc.