@certik 2015-05-02
@isuruf what is the danger? It will raise an exception at runtime. I think that is fine. I think it's better to have less code.