@isuruf 2015-05-02
No danger. We could have a TODO comment for the classes to be implemented.
I implemented most of the remaining functions in eval_double. I'll send a PR, once this issue is merged or closed.