@certik 2015-05-05
So with gcc 4.9.2 on my computer, master is:
```
certik@redhawk:~/repos/csympy(master)$ benchmarks/eval_double1 
4441ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/csympy(master)$ benchmarks/eval_double1 
4415ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/csympy(master)$ benchmarks/eval_double1 
4419ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/csympy(master)$ benchmarks/eval_double1 
4431ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/csympy(master)$ benchmarks/eval_double1 
4570ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/csympy(master)$ benchmarks/eval_double1 
4477ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/csympy(master)$ benchmarks/eval_double1 
4448ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
```
this PR is:
```
certik@redhawk:~/repos/csympy(pr-428)$ ./benchmarks/eval_double1 
4758ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/csympy(pr-428)$ ./benchmarks/eval_double1 
4596ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/csympy(pr-428)$ ./benchmarks/eval_double1 
4590ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/csympy(pr-428)$ ./benchmarks/eval_double1 
4588ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/csympy(pr-428)$ ./benchmarks/eval_double1 
4546ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/csympy(pr-428)$ ./benchmarks/eval_double1 
4505ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/csympy(pr-428)$ ./benchmarks/eval_double1 
4602ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
```
And for comparison, enabling the single dispatch:
```diff
diff --git a/src/eval_double.cpp b/src/eval_double.cpp
index ce05bef..e3c0d29 100644
--- a/src/eval_double.cpp
+++ b/src/eval_double.cpp
@@ -23,7 +23,7 @@ namespace CSymPy {
  * default. To instead use single dispatch, uncomment the
  * CSYMPY_SINGLE_DISPATCH define below.
  */
-// #define CSYMPY_SINGLE_DISPATCH
+#define CSYMPY_SINGLE_DISPATCH
 
 class EvalDoubleVisitor : public BaseVisitor<EvalDoubleVisitor> {
 private:
```
one gets:
```
certik@redhawk:~/repos/csympy(pr-428)$ ./benchmarks/eval_double1 
4621ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/csympy(pr-428)$ ./benchmarks/eval_double1 
4905ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/csympy(pr-428)$ ./benchmarks/eval_double1 
4793ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/csympy(pr-428)$ ./benchmarks/eval_double1 
4583ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/csympy(pr-428)$ ./benchmarks/eval_double1 
4852ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/csympy(pr-428)$ ./benchmarks/eval_double1 
4793ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
certik@redhawk:~/repos/csympy(pr-428)$ ./benchmarks/eval_double1 
4909ms
r (double) = 9.85647
r (exact)  = 9.85647
error      = 5.32907e-15
```
So using the fastest times in each, this PR is about 2% slower than master, and the single dispatch is about 4% slower.

I am not against merging it. But overall, I really like the single dispatch, especially I think it can get faster by storing the class type as a variable, instead of a virtual function as we do know (I would really like to try it out soon, so that we get an idea about the possible speedup). Few percent slowdown I think is fine.

This PR is my second preference. The original visitor pattern is third, though it is the fastest.