@certik 2015-05-05
Anyway, to move forward, I am +1 to merge it, we can currently use any of the three approaches in the code.