@isuruf 2015-05-06
Thanks for the review. Storing the type code should give a speedup for single dispatch. Let's look at how it affects the other benchmarks.